requirements.txt
# Pin Werkzeug because it keeps breaking Flask!
Werkzeug==2.1.2

# Runtime dependencies
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
psycopg2
python-dotenv==0.20.0

# Runtime tools
gunicorn==20.1.0
honcho==1.1.0

# Code quality
pylint==2.14.0
flake8==4.0.1
black==22.3.0

# Testing dependencies
nose==1.3.7
pinocchio==0.4.3
factory-boy==2.12.0

# Code Coverage
coverage==6.3.2
codecov==2.1.12

# Utilities
httpie==3.2.1

#Flask talisman
Flask-Talisman==1.0.0
Flask-Cors
boto3